Add initial script for parsing event history from pdf
Glow only exports individual events (feedings, diaper changes) in the form of a calendar graphic. This commit adds an initial pass at a script for extracting events from this graphic. The PDF is first converted to a jpeg and then processed by hard-coding assumptions about how the data is laid out at a pixel level.